Reworked FDK-AAC:
* Added custom title input
* Added custom command line
* Resized window
* Added view command button
* Adjusted command line viewer for main gui
* Adjusted start job command line

Bugfix for 'View Commands' inside of the codec window. If you clicked the 'X' on the top right hand corner of the screen, it would not allow the window to come back up.

I resolved that for every codec that I've re-worked so far. Will resolve that for the rest of the codecs as I work down them.