Adjusted gain/dolby pro logic II code
FFMPEG doesn't allow multiple -af filters, so I reworked the code and made some changes.

There is now a function that combines both dolby pro logic II and gain to make a single working command for the -af function.

I have changed gain from a drop down menu with a limit of +- 10 to a spinbox with the ablity to go from -30 to +30.

Everything for this codec is working flawless in this regard.