Opus codec changes
Removed the old gain drop down menu, this is replaced by a spinbox to where you can input it manually up to 30 and -30.

I have also added a Dolby Pro Logic II check button in this codec, this is enabled as long as the audio channels is set to 2/Stereo.

Resized the window, removed some code that was no longer needed, fixed some code in the show command line/start job line.