v1.2
Slightly increased the GUI main window size
The video checkbuttons name has been changed to Best video + Audio Muxed File
The audio checkbutton is now renamed to 'Audio Only'
Added a new audio drop down menu, that gives the user the ability to Extract Only or encode to quality 0 on all supported audio codecs (Quality 0 is the highest quality VBR there is - I'm attempting to keep this GUI simple other wise I could put CBR/VBR and a selection of all qualities. Currently I'll keep it this way...simple/highest quality)
Added the option 'noplaylist' and set it to True, when pasting links that are inside of a playlist. This fixed a bug when scraping the link for the title input for the program (The simple version of this program will not deal with playlist at all)
Added the code to always prefer ffmpeg